+++ package.json
@@ -34,6 +34,7 @@
     "url-join": "^4.0.1",
     "use-places-autocomplete": "^1.9.3",
     "web-vitals": "^1.0.1",
+    "xlsx": "^0.17.1",
     "yup": "^0.32.9"
   },
   "scripts": {
